+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/spi/v1/GapicSpannerRpc.java
@@ -280,6 +280,8 @@ public class GapicSpannerRpc implements SpannerRpc {
 
   private Supplier<Boolean> directPathEnabledSupplier = () -> false;
 
+  private final GrpcCallContext baseGrpcCallContext;
+
   public static GapicSpannerRpc create(SpannerOptions options) {
     return new GapicSpannerRpc(options);
   }
@@ -333,6 +335,7 @@ public class GapicSpannerRpc implements SpannerRpc {
     this.endToEndTracingEnabled = options.isEndToEndTracingEnabled();
     this.numChannels = options.getNumChannels();
     this.isGrpcGcpExtensionEnabled = options.isGrpcGcpExtensionEnabled();
+    this.baseGrpcCallContext = createBaseCallContext();
 
     if (initializeStubs) {
       // First check if SpannerOptions provides a TransportChannelProvider. Create one
@@ -370,6 +373,9 @@ public class GapicSpannerRpc implements SpannerRpc {
         defaultChannelProviderBuilder.setAttemptDirectPath(true);
         defaultChannelProviderBuilder.setAttemptDirectPathXds();
       }
+
+      options.enablegRPCMetrics(defaultChannelProviderBuilder);
+
       if (options.isUseVirtualThreads()) {
         ExecutorService executor =
             tryCreateVirtualThreadPerTaskExecutor("spanner-virtual-grpc-executor");
@@ -681,6 +687,10 @@ public class GapicSpannerRpc implements SpannerRpc {
         && options.getHost() != null
         && options.getHost().startsWith("http://localhost")
         && options.getHost().endsWith(emulatorHost);
+  }
+
+  public static boolean isEnableAFEServerTiming() {
+    return "false".equalsIgnoreCase(System.getenv("SPANNER_DISABLE_AFE_SERVER_TIMING"));
   }
 
   private static final RetrySettings ADMIN_REQUESTS_LIMIT_EXCEEDED_RETRY_SETTINGS =
@@ -1974,8 +1984,25 @@ public class GapicSpannerRpc implements SpannerRpc {
       future.cancel(true);
       throw SpannerExceptionFactory.propagateInterrupt(e);
     } catch (Exception e) {
-      throw newSpannerException(context, e);
+      throw newSpannerException(context, e, null);
     }
+  }
+
+  private GrpcCallContext createBaseCallContext() {
+    GrpcCallContext context = GrpcCallContext.createDefault();
+    if (compressorName != null) {
+      // This sets the compressor for Client -> Server.
+      context = context.withCallOptions(context.getCallOptions().withCompression(compressorName));
+    }
+    if (endToEndTracingEnabled) {
+      context = context.withExtraHeaders(metadataProvider.newEndToEndTracingHeader());
+    }
+    if (isEnableAFEServerTiming()) {
+      context = context.withExtraHeaders(metadataProvider.newAfeServerTimingHeader());
+    }
+    return context
+        .withStreamWaitTimeoutDuration(waitTimeout)
+        .withStreamIdleTimeoutDuration(idleTimeout);
   }
 
   // Before removing this method, please verify with a code owner that it is not used
@@ -2002,7 +2029,7 @@ public class GapicSpannerRpc implements SpannerRpc {
       ReqT request,
       MethodDescriptor<ReqT, RespT> method,
       boolean routeToLeader) {
-    GrpcCallContext context = GrpcCallContext.createDefault();
+    GrpcCallContext context = this.baseGrpcCallContext;
     if (options != null) {
       if (this.isGrpcGcpExtensionEnabled) {
         // Set channel affinity in gRPC-GCP.
@@ -2019,16 +2046,9 @@ public class GapicSpannerRpc implements SpannerRpc {
         context = context.withChannelAffinity(Option.CHANNEL_HINT.getLong(options).intValue());
       }
     }
-    if (compressorName != null) {
-      // This sets the compressor for Client -> Server.
-      context = context.withCallOptions(context.getCallOptions().withCompression(compressorName));
-    }
     context = context.withExtraHeaders(metadataProvider.newExtraHeaders(resource, projectName));
     if (routeToLeader && leaderAwareRoutingEnabled) {
       context = context.withExtraHeaders(metadataProvider.newRouteToLeaderHeader());
-    }
-    if (endToEndTracingEnabled) {
-      context = context.withExtraHeaders(metadataProvider.newEndToEndTracingHeader());
     }
     if (callCredentialsProvider != null) {
       CallCredentials callCredentials = callCredentialsProvider.getCallCredentials();
@@ -2037,10 +2057,6 @@ public class GapicSpannerRpc implements SpannerRpc {
             context.withCallOptions(context.getCallOptions().withCallCredentials(callCredentials));
       }
     }
-    context =
-        context
-            .withStreamWaitTimeoutDuration(waitTimeout)
-            .withStreamIdleTimeoutDuration(idleTimeout);
     CallContextConfigurator configurator = SpannerOptions.CALL_CONTEXT_CONFIGURATOR_KEY.get();
     ApiCallContext apiCallContextFromContext = null;
     if (configurator != null) {