+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/it/ITQueryTest.java
@@ -268,9 +268,6 @@ public class ITQueryTest {
 
   @Test
   public void bindFloat32() {
-    assumeFalse("Emulator does not support FLOAT32 yet", isUsingEmulator());
-    assumeTrue("FLOAT32 is currently only supported in cloud-devel", isUsingCloudDevel());
-
     Struct row =
         execute(Statement.newBuilder(selectValueQuery).bind("p1").to(2.0f), Type.float32());
     assertThat(row.isNull(0)).isFalse();
@@ -279,9 +276,6 @@ public class ITQueryTest {
 
   @Test
   public void bindFloat32Null() {
-    assumeFalse("Emulator does not support FLOAT32 yet", isUsingEmulator());
-    assumeTrue("FLOAT32 is currently only supported in cloud-devel", isUsingCloudDevel());
-
     Struct row =
         execute(Statement.newBuilder(selectValueQuery).bind("p1").to((Float) null), Type.float32());
     assertThat(row.isNull(0)).isTrue();
@@ -343,7 +337,6 @@ public class ITQueryTest {
   @Test
   public void bindJson() {
     assumeFalse("JSON are not supported on POSTGRESQL", dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support JSON", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -357,7 +350,6 @@ public class ITQueryTest {
   @Test
   public void bindJsonEmpty() {
     assumeFalse("JSON are not supported on POSTGRESQL", dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support JSON", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").to(Value.json("{}")), Type.json());
@@ -368,7 +360,6 @@ public class ITQueryTest {
   @Test
   public void bindJsonNull() {
     assumeFalse("JSON is not supported on POSTGRESQL", dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support JSON", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").to(Value.json(null)), Type.json());
@@ -452,8 +443,6 @@ public class ITQueryTest {
 
   @Test
   public void bindInterval() {
-    assumeFalse(
-        "INTERVAL is not yet supported on Emulator", EmulatorSpannerHelper.isUsingEmulator());
     Interval d = Interval.parseFromString("P1Y2M3DT4H5M6.789123S");
     Struct row = execute(Statement.newBuilder(selectValueQuery).bind("p1").to(d), Type.interval());
     assertThat(row.isNull(0)).isFalse();
@@ -462,8 +451,6 @@ public class ITQueryTest {
 
   @Test
   public void bindIntervalNull() {
-    assumeFalse(
-        "INTERVAL is not yet supported on Emulator", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").to((Interval) null), Type.interval());
@@ -472,7 +459,6 @@ public class ITQueryTest {
 
   @Test
   public void bindNumeric() {
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     BigDecimal b = new BigDecimal("1.1");
     Statement.Builder statement = Statement.newBuilder(selectValueQuery);
     Type expectedType = Type.numeric();
@@ -497,7 +483,6 @@ public class ITQueryTest {
 
   @Test
   public void bindNumericNull() {
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     Statement.Builder statement = Statement.newBuilder(selectValueQuery);
     Type expectedType = Type.numeric();
     if (dialect.dialect == Dialect.POSTGRESQL) {
@@ -512,7 +497,6 @@ public class ITQueryTest {
 
   @Test
   public void bindNumeric_doesNotPreservePrecision() {
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     BigDecimal b = new BigDecimal("1.10");
     Statement.Builder statement = Statement.newBuilder(selectValueQuery);
     Type expectedType = Type.numeric();
@@ -600,9 +584,6 @@ public class ITQueryTest {
 
   @Test
   public void bindFloat32Array() {
-    assumeFalse("Emulator does not support FLOAT32 yet", isUsingEmulator());
-    assumeTrue("FLOAT32 is currently only supported in cloud-devel", isUsingCloudDevel());
-
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -625,9 +606,6 @@ public class ITQueryTest {
 
   @Test
   public void bindFloat32ArrayEmpty() {
-    assumeFalse("Emulator does not support FLOAT32 yet", isUsingEmulator());
-    assumeTrue("FLOAT32 is currently only supported in cloud-devel", isUsingCloudDevel());
-
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -640,9 +618,6 @@ public class ITQueryTest {
 
   @Test
   public void bindFloat32ArrayNull() {
-    assumeFalse("Emulator does not support FLOAT32 yet", isUsingEmulator());
-    assumeTrue("FLOAT32 is currently only supported in cloud-devel", isUsingCloudDevel());
-
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").toFloat32Array((float[]) null),
@@ -728,7 +703,6 @@ public class ITQueryTest {
   public void bindJsonArray() {
     assumeFalse(
         "array JSON binding is not supported on POSTGRESQL", dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support JSON", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -744,7 +718,6 @@ public class ITQueryTest {
   @Test
   public void bindJsonArrayEmpty() {
     assumeFalse("JSON is not supported on POSTGRESQL", dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support JSON", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").toJsonArray(Collections.emptyList()),
@@ -756,7 +729,6 @@ public class ITQueryTest {
   @Test
   public void bindJsonArrayNull() {
     assumeFalse("JSON is not supported on POSTGRESQL", dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support JSON", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").toJsonArray(null),
@@ -913,8 +885,6 @@ public class ITQueryTest {
 
   @Test
   public void bindIntervalArray() {
-    assumeFalse(
-        "INTERVAL is not yet supported on Emulator", EmulatorSpannerHelper.isUsingEmulator());
     Interval d1 = Interval.parseFromString("P-1Y-2M-3DT4H5M6.789123S");
     Interval d2 = Interval.parseFromString("P1Y2M3DT-4H-5M-6.789123S");
     Struct row =
@@ -927,8 +897,6 @@ public class ITQueryTest {
 
   @Test
   public void bindIntervalArrayEmpty() {
-    assumeFalse(
-        "INTERVAL is not yet supported on Emulator", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -941,8 +909,6 @@ public class ITQueryTest {
 
   @Test
   public void bindIntervalArrayNull() {
-    assumeFalse(
-        "INTERVAL is not yet supported on Emulator", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery).bind("p1").toIntervalArray(null),
@@ -953,7 +919,6 @@ public class ITQueryTest {
   @Test
   public void bindNumericArrayGoogleStandardSQL() {
     assumeTrue(dialect.dialect == Dialect.GOOGLE_STANDARD_SQL);
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     BigDecimal b1 = new BigDecimal("3.14");
     BigDecimal b2 = new BigDecimal("6.626");
 
@@ -968,7 +933,6 @@ public class ITQueryTest {
   @Test
   public void bindNumericArrayPostgreSQL() {
     assumeTrue(dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -982,7 +946,6 @@ public class ITQueryTest {
   @Test
   public void bindNumericArrayEmptyGoogleStandardSQL() {
     assumeTrue(dialect.dialect == Dialect.GOOGLE_STANDARD_SQL);
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)
@@ -996,7 +959,6 @@ public class ITQueryTest {
   @Test
   public void bindNumericArrayEmptyPostgreSQL() {
     assumeTrue(dialect.dialect == Dialect.POSTGRESQL);
-    assumeFalse("Emulator does not yet support NUMERIC", EmulatorSpannerHelper.isUsingEmulator());
     Struct row =
         execute(
             Statement.newBuilder(selectValueQuery)